Alternative Score Card
Added an alternative score card component, which displays the score as a
series of dots. As part of this change I added new mutations and states.
Notably, I changed the scoring mutation and score state. The score state
is now an object wherein each player has a score object. This simplifies
the scoring mutation process. This change impacts several components. It
also obviates the need for a separate Score component (slated to be
removed).

Also adjusted the positioning of the player turn display in multiplayer
mode. The logic for this has been taken out of the Score component and
placed into the GameBoard component.